# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Spotifeel",
    "description": "API for retrieving playlists based on emotions"
  },
  "servers": [
    {
      "description": "Local server",
      "url": "http://localhost:8888"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
    "/feelings": {
      "get": {
        "summary": "Retrieve all emotions",
        "description": "Retrieves a list of all emotions supported by the API",
        "responses": {
          "200": {
            "description": "List of all emotions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Emotion"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/feelings/{emotionId}": {
      "get": {
        "summary": "Retrieve emotion by ID",
        "description": "Retrieves a specific emotion based on its ID",
        "parameters": [
          {
            "name": "emotionId",
            "in": "path",
            "required": true,
            "description": "ID for emotions: 1=furious, 2=frustrated, 3=horrified, 4=disappointed, 5=euphoric, 6=loving, 7=happy, 8=useless, 9=regretful, 10=dejected, 11=unhappy, 12=scared, 13=anxious",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 13
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Emotion found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Emotion"
                }
              }
            }
          },
          "404": {
            "description": "Emotion with the specified ID does not exist"
          }
        }
      }
    },
    "/playlists/{emotionId}": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create a playlist based on emotion",
        "description": "Creates a new playlist based on the user's emotion and returns the created playlist",
        "parameters": [
          {
            "name": "emotionId",
            "in": "path",
            "required": true,
            "description": "ID for emotions: 1=furious, 2=frustrated, 3=horrified, 4=disappointed, 5=euphoric, 6=loving, 7=happy, 8=useless, 9=regretful, 10=dejected, 11=unhappy, 12=scared, 13=anxious",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 13
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Playlist created and found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Playlist"
                }
              }
            }
          },
          "401": {
            "description": "User must be logged in to create a playlist"
          },
          "404": {
            "description": "Emotion with the specified ID does not exist"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Emotion": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "Playlist": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "songs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Song"
            }
          },
          "spotifyUrl": {
            "type": "string",
            "description": "URL to the playlist on Spotify"
          }
        }
      },
      "Song": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      }
    }
  }
}